requirements.txt
alembic==1.6.5
autopep8==1.5.7
cachelib==0.3.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
ecdsa==0.17.0
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==19.9.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycodestyle==2.7.0
python-dateutil==2.8.2
python-editor==1.0.4
python-jose==3.3.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.23
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
